Add interface and test implementation for BuildStorageManager
